pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminalbot"
version = "0.1.0"
description = "AI-powered terminal troubleshooting assistant for Fedora Linux"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TerminalBot Contributors"}
]
keywords = ["linux", "troubleshooting", "cli", "ai", "fedora"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

# Core dependencies - minimal, always installed (~15MB total)
dependencies = [
    "typer>=0.9.0",        # CLI framework (modern, type-based)
    "rich>=13.7.0",        # Terminal formatting (~5MB)
    "psutil>=5.9.0",       # System utilities (~1MB)
    "pyyaml>=6.0",         # Config parsing (~1MB)
    "pydantic>=2.5.0",     # Data validation (~3MB)
    "pydantic-settings>=2.0.0",  # Settings management
    "httpx>=0.27.0",       # HTTP client (~3MB, lighter than requests)
]

[project.optional-dependencies]
# LLM features - install only if using AI capabilities
llm = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
]

# Local LLM support via Ollama
ollama = [
    "ollama>=0.1.0",
]

# Cloud LLM providers
cloud = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
]

# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

# Full installation (all features)
full = [
    "terminalbot[llm,ollama,cloud]",
]

[project.scripts]
terminalbot = "terminalbot.cli.main:cli"

[project.urls]
Homepage = "https://github.com/terminalbot/terminalbot"
Repository = "https://github.com/terminalbot/terminalbot"
Issues = "https://github.com/terminalbot/terminalbot/issues"

[tool.setuptools]
packages = ["terminalbot"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
terminalbot = ["config/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=terminalbot --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
